requirements.txt
#Django==5.0.1
#Pillow==10.2.0
#django-environ==0.11.2
#django-crispy-forms==2.1
#crispy-tailwind==1.0.3
#django-compressor==4.4
#whitenoise==6.6.0
#gunicorn==21.2.0
#psycopg2-binary==2.9.9
#django-ckeditor==6.7.0
#django-taggit==5.0.1
#django-analytics==0.0.1
#django-cors-headers==4.3.1
#celery==5.3.6
#redis==5.0.1
#django-redis==5.4.0
#django-storages==1.14.2
#boto3==1.34.34
#python-dotenv==1.0.0


# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter
dj-database-url>=2.1.0  # Database configuration from URL

# Image Processing
Pillow>=10.0.0

# Rich Text Editor
django-ckeditor>=6.7.0

# Tags
django-taggit>=5.0.0

# Forms
django-crispy-forms>=2.1
crispy-tailwind>=1.0.3

# Environment Variables
python-decouple>=3.8

# Email
django-anymail>=10.2  # For advanced email providers (optional)

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Security
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0  # Rate limiting

# Performance & Caching
django-redis>=5.4.0  # Redis cache backend
redis>=5.0.0

# Monitoring & Logging (optional but recommended)
sentry-sdk>=1.39.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
factory-boy>=3.3.0
faker>=20.0.0

# Development Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
ipython>=8.18.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0  # Static file serving

# File Storage (optional - for cloud storage)
django-storages>=1.14.2
boto3>=1.34.0  # For AWS S3

# SEO
django-meta>=2.3.0
django-robots>=6.1

# Analytics (optional)
django-analytical>=3.1.0

# Admin Enhancements
django-admin-interface>=0.27.0
django-grappelli>=3.0.8

# API Documentation (if building API)
drf-yasg>=1.21.7

# Task Queue (optional - for background tasks)
celery>=5.3.4
django-celery-beat>=2.5.0
flower>=2.0.1  # Celery monitoring

# Search (optional - for advanced search)
django-haystack>=3.2.1
elasticsearch>=8.11.0

# Social Auth (optional)
social-auth-app-django>=5.4.0

# Export/Import
django-import-export>=3.3.5

# Sitemap
django.contrib.sitemaps  # Built-in

# Comments (alternative to custom implementation)
django-contrib-comments>=2.2.0

# Markdown Support (optional)
markdown>=3.5.1
django-markdownx>=4.0.2

# Translation (if multilingual support needed)
django-modeltranslation>=0.18.11
django-parler>=2.3

# Webhooks (optional)
django-webhook>=1.3.0

# Frontend Asset Management
django-compressor>=4.4  # CSS/JS compression
django-webpack-loader>=2.0.1  # If using Webpack

# Health Checks
django-health-check>=3.18.1

# Backup
django-dbbackup>=4.0.2